refactor: use identical code to handle chunked uploads of binary CoValue.
