+++ experiments/servers/package.json
@@ -28,7 +28,6 @@
     "body-parser": "^1.20.2",
     "express": "^4.19.2",
     "finalhandler": "^1.2.0",
-    "multer": "1.4.5-lts.1",
     "serve-static": "^1.15.0",
     "spdy": "^4.0.2",
     "uuid": "^10.0.0",
@@ -41,7 +40,6 @@
     "@types/express": "^4.17.21",
     "@types/finalhandler": "^1.2.3",
     "@types/mime-types": "^2.1.4",
-    "@types/multer": "^1.4.12",
     "@types/node": "^22.5.0",
     "@types/serve-static": "^1.15.7",
     "@types/spdy": "^3.4.9",
